Create Limited Data Transmitter.cfg
I like the look of R&D_node_icon_advunmanned vs. DataTransmitter because it's darker, so it fits better with the other tech-tree icons that are used as module symbols (vs. the lighter part-category icons that DataTransmitter and the other part icons are meant to fit with).  On the other hand, data transmitters can be put on things besides satellites, so the satellite image may be a bit misleading.  What do you think?